requirements.txt
streamlit
pandas
requests
beautifulsoup4
plotly
google-generativeai
